add new entities to models package
