+++ entity/models/comment_rating.go
@@ -1,4 +1,4 @@
-package entity
+package models
 
 import (
 	"github.com/google/uuid"
